Implement Bosch BSEC library as a Mynewt package
This the first main commit. The Bosch BSEC library is implemented as a Mynewt
package that can be downloaded and included in any Mynewt project. This package
does not included the closed-source binary BSEC blob. This can be downloaded
from the Bosch website at https://www.bosch-sensortec.com/bst/products/all_products/bsec.
The .a file needs to be placed in order package /src directory in order for it to be
linked with the project.